Add package with layers and losses
